# apd: avoid function calls on hot paths