--- a/table.go
+++ b/table.go
@@ -52,6 +52,7 @@ func init() {
 }
 
 // NumDigits returns the number of decimal digits of d.Coeff.
+//gcassert:inline
 func (d *Decimal) NumDigits() int64 {
 	return NumDigits(&d.Coeff)
 }